fix(apc_emit): pooled pages hold pre-edit text
The editor pool and the minimap thumb keyed on the display snapshot
version, which counts fold toggles alone. A file outside git with no
diagnostics moved nothing else in the page hash. An edit therefore left
every buffered page stale, and a wheel glide composited the text from
before it.

Both now key on the buffer version paired with the full paint version,
the pair the pane replay cache already used. Every part of it moves only
on work that relaid the rows. The pages refill when the text changes and
stay cached when it does not.

+++ widgets/src/pool.rs
@@ -73,12 +73,12 @@ struct MinimapView {
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct MinimapWindowInputs {
     scroll_offset_bits: u32,
-    map_version: usize,
+    map_version: u64,
     viewport_rows: u32,
 }
 
 impl MinimapWindowInputs {
-    pub fn new(scroll_offset: f32, map_version: usize, viewport_rows: u32) -> Self {
+    pub fn new(scroll_offset: f32, map_version: u64, viewport_rows: u32) -> Self {
         Self {
             scroll_offset_bits: scroll_offset.to_bits(),
             map_version,
